requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# AI Services
anthropic==0.8.1
openai==1.8.0
httpx==0.26.0
aiohttp==3.9.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyclerkapi==1.0.0

# Task Queue
celery[redis]==5.3.4
redis==5.0.1
flower==2.0.1

# Notifications
resend==0.7.2
twilio==8.11.0

# Monitoring
sentry-sdk[fastapi]==1.39.1
logtail-python==0.3.1
posthog==3.1.0

# Utils
python-dateutil==2.8.2
pytz==2023.3
numpy==1.24.3
pandas==2.0.3

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.26.0
faker==22.0.0